Removed constant bool fields with all occurences.
