--- src/System.Net.Http/tests/FunctionalTests/HttpClientHandlerTest.ServerCertificates.Windows.cs
@@ -1,11 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace System.Net.Http.Functional.Tests
-{
-    public abstract partial class HttpClientHandler_ServerCertificates_Test
-    {
-        internal bool BackendSupportsCustomCertificateHandling => true;
-    }
-}
